\name{testRscalaPackage-package}
\alias{testRscalaPackage-package}
\alias{testRscalaPackage}
\docType{package}
\title{
\packageTitle{testRscalaPackage}
}
\description{
\packageDescription{testRscalaPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testRscalaPackage}
\packageIndices{testRscalaPackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{testRscalaPackage}

Maintainer: \packageMaintainer{testRscalaPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
